refactor(FoodResponseDTO): Updated price attribute

+++ src/main/java/com/io/github/pedroolivsz/cardapio/DTO/FoodResponseDTO.java
@@ -2,9 +2,7 @@ package com.io.github.pedroolivsz.cardapio.DTO;
 
 import com.io.github.pedroolivsz.cardapio.entity.Food;
 
-import java.math.BigDecimal;
-
-public record FoodResponseDTO(Long id, String title, String image, BigDecimal price) {
+public record FoodResponseDTO(Long id, String title, String image, Integer price) {
     public FoodResponseDTO(Food food) {
         this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
     }
